qa-agent/requirements.txt
anthropic>=0.40.0
PyGithub>=2.1.1
